feat: Part 1. updated_instance function

+++ crates/engine-core/src/lib.rs
@@ -23,16 +23,16 @@ impl AetherCore {
 
 #[derive(Component, Default)]
 pub struct Transform {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 #[derive(Component, Default)]
 pub struct Velocity {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 #[derive(Bundle)]
